feat(core): make the AvroSchema model hot-path ready
Before any encoder or decoder reads AvroSchema (M3-09/M3-11), remove the
shapes that would be slow per value:

- AvroSchema.type is a stored field of the sealed base, no longer an
  abstract getter overridden by 14 subclasses.
- AvroSchema.logicalTypeName is computed once at construction (null for
  unions); it replaces the ResolvedSchema extension that looked up props
  and cast JSON on every call.
- EnumSchema.getSymbolIndexOrNull: O(1) symbol -> index, built once.
- RecordSchema.getFieldIndexOrNull / getFieldOrNull by name or alias:
  index built lazily after the LockableList is locked, published through
  a @Volatile field (safe on JVM and native), throws if queried before.
- FixedSchema.size is Int (non-negative), no longer UInt with mangled
  JVM names.

Cached state stays out of equals/hashCode/toString; B1's cycle-safe
equality is unchanged. Findings in docs/plans/notes/m3-04.md.

+++ apache-interop/src/main/kotlin/com/github/avrokotlin/avro4k/AvroSchemaApacheInterop.kt
@@ -109,7 +109,7 @@ private fun from(schema: Schema, seenNamedTypes: MutableMap<String, NamedSchema>
         Schema.Type.FIXED ->
             FixedSchema(
                 name = Name(schema.name, schema.namespace),
-                size = schema.fixedSize.toUInt(),
+                size = schema.fixedSize,
                 doc = schema.doc,
                 aliases = schema.aliasesWithSpace,
                 props = schema.objectProps.toJsonElementMap()
